Added CoreLocation to project and implemented Location service. Changed lat/long variables in model to double.

+++ JWeather/JWeather/Entities/ApiResponse.swift
@@ -6,8 +6,8 @@
 //
 
 struct ApiResponse: Decodable {
-    let lat: Float
-    let lon: Float
+    let lat: Double
+    let lon: Double
     let timezone: String
     let timezoneOffset: Int
     let current: Current
